os, syscall: implement Stat and Lstat
File.Stat is left as a stub for now.

Tests are a bit stubbed down because os.ReadDir, os.Symlink, and t.TempDir are not yet (fully) implemented.
TODO: reimport tests from upstream as those materialize.

The definition of Stat_t is somewhat suspect; it is taken from upstream for current MacOS,
but is probably supposed to match the definition in the current libc
(lib/musl/arch/$ARCH or lib/picolibc/newlib/libc/include).

Here be dragons.